pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Compares two configuration files and shows a difference."
authors = ["Alexey Cheremushkin <a.ch3re359@yandex.ru>"]
repository = 'https://github.com/seeu359/python-project-lvl2'
classifiers = ['?']
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff_script:main"
